+++ oshdb-tool/etl/src/main/java/org/heigit/bigspatialdata/oshdb/tool/importer/transform/reader/TransfromNodeReaders.java
@@ -7,9 +7,8 @@ import java.util.List;
 import java.util.PriorityQueue;
 import java.util.Set;
 import java.util.TreeSet;
-
-import org.heigit.bigspatialdata.oshdb.index.zfc.ZGrid;
 import org.heigit.bigspatialdata.oshdb.tool.importer.osh.TransformOSHNode;
+import org.heigit.bigspatialdata.oshdb.tool.importer.util.ZGrid;
 
 public class TransfromNodeReaders {
   private static class TransformNodeReader extends TransformReader<TransformOSHNode> {
